+++ server/src/ast/dtc/comment.ts
@@ -15,12 +15,15 @@
  */
 
 import { ASTBase } from "../base";
-import { TokenIndexes } from "../../types";
+import { Token, TokenIndexes } from "../../types";
 
 export class CommentBlock extends ASTBase {
   constructor(readonly comments: Comment[]) {
     super();
     comments.forEach(this.addChild.bind(this));
+  }
+  toString(): string {
+    return this.comments.map((c) => c.toString()).join("\n");
   }
 }
 
@@ -29,5 +32,29 @@ export class Comment extends ASTBase {
     super(tokenIndexes);
     this.semanticTokenType = "comment";
     this.semanticTokenModifiers = "documentation";
+  }
+
+  toString(): string {
+    let prev: Token | undefined;
+    let token: Token | undefined = this.firstToken;
+    let str = "";
+    while (token !== this.lastToken) {
+      str += token?.value.padStart(
+        token.value.length + (prev ? token.pos.col - prev.pos.colEnd : 0),
+        " "
+      );
+      prev = token;
+      token = token?.nextToken;
+    }
+
+    if (token) {
+      str += token?.value.padStart(
+        token.value.length + (prev ? token.pos.col - prev.pos.colEnd : 0),
+        " "
+      );
+    }
+
+    const match = str.match(/^\s*(?:\/\/|\/\*+)\s*(.*?)\s*\*?\/?\s*$/s);
+    return match ? match[1] : "";
   }
 }