+++ server/src/getDocumentFormatting.ts
@@ -170,8 +170,82 @@ export async function getDocumentFormatting(
     result.push(...removeTrailingWhitespace(splitDocument, result));
   }
 
-  return result;
+  const formatOnOffMeta = pairFormatOnOff(fileRootAsts, splitDocument);
+  return formatOnOffMeta.length
+    ? result.filter(
+        (edit) => !isFormattingDisabledAt(edit.range.start, formatOnOffMeta)
+      )
+    : result;
 }
+const pairFormatOnOff = (
+  fileRootAsts: ASTBase[],
+  documentLines: string[]
+): Range[] => {
+  const last = Position.create(
+    documentLines.length - 1,
+    documentLines.at(-1)?.length ?? 0
+  );
+
+  const formatControlRanges: Range[] = [];
+  let pendingOff: { start: Position } | undefined;
+
+  const controlComments = fileRootAsts
+    .filter(
+      (ast) =>
+        (ast instanceof CommentBlock || ast instanceof Comment) &&
+        /^dts-format (on|off)$/.test(ast.toString().trim())
+    )
+
+    .sort((a, b) => a.firstToken.pos.line - b.firstToken.pos.line);
+
+  controlComments.forEach((ast) => {
+    const value = ast.toString().trim();
+
+    if (value === "dts-format off") {
+      pendingOff = {
+        start: Position.create(
+          ast.firstToken.pos.line,
+          ast instanceof CommentBlock ? ast.firstToken.pos.colEnd : 0
+        ),
+      };
+    } else if (value === "dts-format on" && pendingOff) {
+      const end = Position.create(
+        ast.lastToken.pos.line,
+        ast instanceof CommentBlock
+          ? ast.lastToken.pos.colEnd - 1
+          : documentLines[ast.lastToken.pos.line - 1].length
+      );
+      formatControlRanges.push(Range.create(pendingOff.start, end));
+      pendingOff = undefined;
+    }
+  });
+
+  // If still "off" with no "on", use last known AST node as document end
+  if (pendingOff) {
+    formatControlRanges.push(Range.create(pendingOff.start, last));
+  }
+
+  return formatControlRanges;
+};
+
+function comparePositions(a: Position, b: Position): number {
+  if (a.line < b.line) return -1;
+  if (a.line > b.line) return 1;
+  if (a.character < b.character) return -1;
+  if (a.character > b.character) return 1;
+  return 0;
+}
+
+const isFormattingDisabledAt = (
+  pos: Position,
+  disabledRanges: Range[]
+): boolean => {
+  return disabledRanges.some(
+    (range) =>
+      comparePositions(pos, range.start) >= 0 &&
+      comparePositions(pos, range.end) <= 0
+  );
+};
 
 const removeTrailingWhitespace = (
   documentText: string[],